backend/requirements.txt
psycopg2-binary==2.9.11
SQLAlchemy==2.0.21
python-dotenv==1.0.1
fastapi>=0.95
uvicorn>=0.23
pytest>=8.0
httpx>=0.24
starlette>=0.28
stripe >= 5.0.0